[TS - Adapter pattern]: rename JSON into json
