+++ typescript/adapter/exercises/json-report.ts
